requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
peft>=0.9.0
einops>=0.7.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.8.0
Pillow>=10.0.0
tqdm>=4.66.0
scipy>=1.11.0
scikit-learn>=1.3.0
gdown>=5.1.0
huggingface_hub>=0.20.0
